+++ react-rss/src/pages/HomePage/hooks/useHomePage.ts
@@ -1,8 +1,5 @@
 import { useEffect, useRef, useState } from 'react'
 
-import { Сharacter, getCharacter } from '../../../api'
-import { useModal } from '../../../components/Modal'
-
 import { useFetchCharacters } from './useFetchCharacters'
 
 export const useHomePage = () => {
@@ -42,26 +39,4 @@ export const useHomePage = () => {
   }
 }
 
-export const useFetchCharacter = () => {
-  const {
-    visible: modalVisible,
-    open: openModal,
-    close: closeModal,
-  } = useModal()
 
-  const [selectCharacter, setSelectCharacter] = useState<Сharacter | null>(null)
-
-  const onCardClick = async (id: number) => {
-    const characterInfo = await getCharacter(id)
-    setSelectCharacter(characterInfo)
-
-    openModal()
-  }
-
-  return {
-    onCardClick,
-    selectCharacter,
-    modalVisible,
-    closeModal,
-  }
-}
